moved implementation of env to a separate file
